cpu: x86_64: fixup interrupts service routine stub

+++ include/cpu/isr.h
@@ -9,6 +9,6 @@ struct isr_frame;
 typedef status_t (*isr_t)(struct isr_frame*);
 
 void isr_init(void);
-void isr_register(int irq, isr_t handler, isr_t* old_handler);
+void isr_register(size_t irq, isr_t handler, isr_t* old_handler);
 
 #endif
